Move AppBrandDesignUpdatePreWarmObserver to design system module
