+++ android-design-system/design-system/src/main/java/com/duckduckgo/common/ui/store/AppBrandDesignUpdatePreWarmObserver.kt
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.app.branddesignupdate
+package com.duckduckgo.common.ui.store
 
 import androidx.lifecycle.LifecycleOwner
 import com.duckduckgo.app.di.AppCoroutineScope
 import com.duckduckgo.app.lifecycle.MainProcessLifecycleObserver
-import com.duckduckgo.common.ui.store.AppBrandDesignUpdateToggles
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
 import com.squareup.anvil.annotations.ContributesMultibinding
